restructure, matrix iterators
