+++ bson_benchmark_test.go
@@ -134,7 +134,7 @@ func BenchmarkDocumentCreation(b *testing.B) {
 		b.Run(test.Name, func(b *testing.B) {
 			for n := 0; n < b.N; n++ {
 				for i := 0; i < test.Samples; i++ {
-					doc, _ = rehydrateDocument(test.Reference, i, test.Metrics, 0)
+					doc, _ = restoreDocument(test.Reference, i, test.Metrics, 0)
 					require.NotNil(b, doc)
 					require.Equal(b, test.Length, doc.Len())
 				}
