+++ iterator_benchmark_test.go
@@ -55,26 +55,6 @@ func BenchmarkIterator(b *testing.B) {
 					}
 				})
 			})
-			b.Run("Matrix", func(b *testing.B) {
-				b.Run("Resolving", func(b *testing.B) {
-					iter := ReadMatrix(ctx, bytes.NewBuffer(data))
-					b.ResetTimer()
-					for n := 0; n < b.N; n++ {
-						if !iter.Next() {
-							break
-						}
-						require.NotNil(b, iter.Document())
-					}
-				})
-				b.Run("Iterating", func(b *testing.B) {
-					for n := 0; n < b.N; n++ {
-						iter := ReadMatrix(ctx, bytes.NewBuffer(data))
-						for iter.Next() {
-							require.NotNil(b, iter.Document())
-						}
-					}
-				})
-			})
 			b.Run("Series", func(b *testing.B) {
 				b.Run("Resolving", func(b *testing.B) {
 					iter := ReadSeries(ctx, bytes.NewBuffer(data))
@@ -89,6 +69,26 @@ func BenchmarkIterator(b *testing.B) {
 				b.Run("Iterating", func(b *testing.B) {
 					for n := 0; n < b.N; n++ {
 						iter := ReadSeries(ctx, bytes.NewBuffer(data))
+						for iter.Next() {
+							require.NotNil(b, iter.Document())
+						}
+					}
+				})
+			})
+			b.Run("Matrix", func(b *testing.B) {
+				b.Run("Resolving", func(b *testing.B) {
+					iter := ReadMatrix(ctx, bytes.NewBuffer(data))
+					b.ResetTimer()
+					for n := 0; n < b.N; n++ {
+						if !iter.Next() {
+							break
+						}
+						require.NotNil(b, iter.Document())
+					}
+				})
+				b.Run("Iterating", func(b *testing.B) {
+					for n := 0; n < b.N; n++ {
+						iter := ReadMatrix(ctx, bytes.NewBuffer(data))
 						for iter.Next() {
 							require.NotNil(b, iter.Document())
 						}
