+++ iterator_sample.go
@@ -24,7 +24,7 @@ func (c *Chunk) streamFlattenedDocuments(ctx context.Context) <-chan *bsonx.Docu
 
 			doc := bsonx.MakeDocument(len(c.metrics))
 			for _, m := range c.metrics {
-				elem, ok := rehydrateFlat(m.originalType, m.Key(), m.Values[i])
+				elem, ok := restoreFlat(m.originalType, m.Key(), m.Values[i])
 				if !ok {
 					continue
 				}
@@ -51,7 +51,7 @@ func (c *Chunk) streamDocuments(ctx context.Context) <-chan *bsonx.Document {
 		defer close(out)
 
 		for i := 0; i < c.nPoints; i++ {
-			doc, _ := rehydrateDocument(c.reference, i, c.metrics, 0)
+			doc, _ := restoreDocument(c.reference, i, c.metrics, 0)
 			select {
 			case <-ctx.Done():
 				return
