+++ util.go
@@ -5,7 +5,38 @@ import (
 	"compress/zlib"
 	"encoding/binary"
 	"math"
+	"time"
+
+	"github.com/mongodb/ftdc/bsonx"
+	"github.com/mongodb/grip/message"
+	"github.com/mongodb/mongo-go-driver/bson"
+	"github.com/pkg/errors"
 )
+
+func readDocument(in interface{}) (*bsonx.Document, error) {
+	switch doc := in.(type) {
+	case *bsonx.Document:
+		return doc, nil
+	case []byte:
+		return bsonx.ReadDocument(doc)
+	case bson.Marshaler:
+		data, err := doc.MarshalBSON()
+		if err != nil {
+			return nil, errors.Wrap(err, "problem with unmarshaler")
+		}
+		return bsonx.ReadDocument(data)
+	case map[string]interface{}, map[string]string, map[string]int, map[string]int64, map[string]uint, map[string]uint64:
+		return nil, errors.New("cannot use a map type as an ftdc value")
+	case bson.M, message.Fields:
+		return nil, errors.New("cannot use a custom map type as an ftdc value")
+	default:
+		data, err := bson.Marshal(in)
+		if err != nil {
+			return nil, errors.Wrap(err, "problem with fallback marshaling")
+		}
+		return bsonx.ReadDocument(data)
+	}
+}
 
 func getOffset(count, sample, metric int) int { return metric*count + sample }
 
@@ -62,3 +93,22 @@ func compressBuffer(input []byte) ([]byte, error) {
 
 func normalizeFloat(in float64) int64 { return int64(math.Float64bits(in)) }
 func restoreFloat(in int64) float64   { return math.Float64frombits(uint64(in)) }
+func epochMs(t time.Time) int64       { return t.UnixNano() / 1000000 }
+func timeEpocMs(in int64) time.Time   { return time.Unix(int64(in)/1000, int64(in)%1000*1000000) }
+
+func isNum(num int, val *bsonx.Value) bool {
+	if val == nil {
+		return false
+	}
+
+	switch val.Type() {
+	case bsonx.TypeInt32:
+		return val.Int32() == int32(num)
+	case bsonx.TypeInt64:
+		return val.Int64() == int64(num)
+	case bsonx.TypeDouble:
+		return val.Double() == float64(num)
+	default:
+		return false
+	}
+}
